refactor: web 패키지 분리
