+++ src/main/java/roomescape/web/UserWebController.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.controller;
+package roomescape.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
